--- modules/dxp/apps/osb/osb-client-extension-util/osb-client-extension-util-zendesk/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay OSB Client Extension Utilities Zendesk
-Bundle-SymbolicName: com.liferay.osb.client.extension.util.zendesk
-Bundle-Version: 1.0.0
