+++ modules/dxp/apps/osb/osb-spring-boot/osb-spring-boot-client-zendesk/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay OSB Spring Boot Client Zendesk
+Bundle-SymbolicName: com.liferay.osb.spring.boot.client.zendesk
+Bundle-Version: 1.0.0
