+++ workspaces/liferay-customer-workspace/client-extensions/liferay-customer-etc-spring-boot/src/main/resources/application-default.properties
@@ -4,9 +4,9 @@
 
 liferay.customer.gcs.bucket.name=
 liferay.customer.gcs.service.account.key=
-liferay.customer.zendesk.api.email.address=
-liferay.customer.zendesk.api.token=
-liferay.customer.zendesk.url=
+liferay.osb.spring.boot.client.zendesk.api.email.address=
+liferay.osb.spring.boot.client.zendesk.api.token=
+liferay.osb.spring.boot.client.zendesk.url=
 
 #
 # LXC
